+++ src/Describer/OpenApiPhpDescriber.php
@@ -12,8 +12,8 @@
 namespace Nelmio\ApiDocBundle\Describer;
 
 use Doctrine\Common\Annotations\Reader;
-use Nelmio\ApiDocBundle\Annotation\Operation;
-use Nelmio\ApiDocBundle\Annotation\Security;
+use Nelmio\ApiDocBundle\Attribute\Operation;
+use Nelmio\ApiDocBundle\Attribute\Security;
 use Nelmio\ApiDocBundle\OpenApiPhp\Util;
 use Nelmio\ApiDocBundle\Util\ControllerReflector;
 use Nelmio\ApiDocBundle\Util\SetsContextTrait;