+++ src/ModelDescriber/Annotations/AnnotationsReader.php
@@ -24,6 +24,7 @@ class AnnotationsReader
     private PropertyPhpDocReader $phpDocReader;
     private OpenApiAnnotationsReader $openApiAnnotationsReader;
     private SymfonyConstraintAnnotationReader $symfonyConstraintAnnotationReader;
+    private ReflectionReader $reflectionReader;
 
     /**
      * @param string[] $mediaTypes
@@ -40,12 +41,14 @@ class AnnotationsReader
             $annotationsReader,
             $useValidationGroups
         );
+        $this->reflectionReader = new ReflectionReader();
     }
 
     public function updateDefinition(\ReflectionClass $reflectionClass, OA\Schema $schema): bool
     {
         $this->openApiAnnotationsReader->updateSchema($reflectionClass, $schema);
         $this->symfonyConstraintAnnotationReader->setSchema($schema);
+        $this->reflectionReader->setSchema($schema);
 
         return $this->shouldDescribeModelProperties($schema);
     }
@@ -66,6 +69,7 @@ class AnnotationsReader
     {
         $this->openApiAnnotationsReader->updateProperty($reflection, $property, $serializationGroups);
         $this->phpDocReader->updateProperty($reflection, $property);
+        $this->reflectionReader->updateProperty($reflection, $property);
         $this->symfonyConstraintAnnotationReader->updateProperty($reflection, $property, $serializationGroups);
     }