+++ tests/Functional/Controller/ApiController81.php
@@ -464,7 +464,8 @@ class ApiController81
     #[Route('/inline_path_parameters')]
     #[OA\Response(response: '200', description: '')]
     public function inlinePathParameters(
-        #[OA\PathParameter] string $product_id
+        #[OA\PathParameter]
+        string $product_id
     ) {
     }